refactor(settings): share the composer's model picker with Settings > Default Model
Settings > Default Model still carried the pre-#361 look: grey cards, always
expanded groups, no provider glyphs, no favorites star, and a hand-rolled
search field. Rather than restyle a second copy, lift #361's picker into one
shared `ModelPickerSheet` that both surfaces open through a preset.

`.composer` and `.serverDefault` differ only in titles, the custom entry's
provider requirement, whether the custom entry can be favorited, whether the
Current / Saved Custom groups render, and whether committing dismisses. The
server-saving surface additionally supplies the checkmark rule, the in-flight
row key, a load status, and its save error, so `DefaultModelPickerView` is now
just the loader and server-saver around the shared view. Its `isChecked` rule
and the 12 tests pinned to it are unchanged.

Settings never prefills the custom provider field, so a bare model ID still
saves exactly as it did before.
