+++ HermesMobileTests/ModelPickerSheetTests.swift
@@ -1,0 +1,83 @@
+import XCTest
+@testable import HermesMobile
+
+/// Pins the two rules the shared picker gained when Settings > Default Model
+/// started using it: which row spins during a server save, and what the custom
+/// entry commits under each configuration.
+final class ModelPickerSheetTests: XCTestCase {
+    private func option(_ id: String, provider: String?) -> ModelCatalogOption {
+        ModelCatalogOption(id: id, displayName: id, providerID: provider)
+    }
+
+    // MARK: In-flight row
+
+    func testInFlightRowMatchesTheSavingModelAndProvider() {
+        let key = ModelFavoriteKey(modelID: "gpt-5", providerID: "openai")
+
+        XCTAssertTrue(
+            ModelPickerSheet.isInFlight(option("gpt-5", provider: "openai"), inFlightKey: key)
+        )
+    }
+
+    func testInFlightRowIgnoresTheSameModelIDUnderAnotherProvider() {
+        let key = ModelFavoriteKey(modelID: "gpt-5", providerID: "openai")
+
+        XCTAssertFalse(
+            ModelPickerSheet.isInFlight(option("gpt-5", provider: "azure"), inFlightKey: key)
+        )
+        XCTAssertFalse(
+            ModelPickerSheet.isInFlight(option("gpt-5", provider: nil), inFlightKey: key)
+        )
+    }
+
+    func testNoRowIsInFlightWithoutAKey() {
+        XCTAssertFalse(
+            ModelPickerSheet.isInFlight(option("gpt-5", provider: "openai"), inFlightKey: nil)
+        )
+    }
+
+    // MARK: Custom entry
+
+    func testServerDefaultCommitsABareModelIDWithNoProvider() {
+        let built = ModelPickerSheet.customOption(
+            modelID: "  my-model  ",
+            providerID: "  ",
+            configuration: .serverDefault
+        )
+
+        XCTAssertEqual(built?.id, "my-model")
+        XCTAssertNil(built?.providerID)
+    }
+
+    func testComposerRequiresAProviderForACustomModel() {
+        XCTAssertNil(
+            ModelPickerSheet.customOption(
+                modelID: "my-model",
+                providerID: "",
+                configuration: .composer
+            )
+        )
+    }
+
+    func testCustomProviderIsLowercasedWhenSupplied() {
+        let built = ModelPickerSheet.customOption(
+            modelID: "my-model",
+            providerID: "OpenAI",
+            configuration: .serverDefault
+        )
+
+        XCTAssertEqual(built?.providerID, "openai")
+    }
+
+    func testEmptyModelIDNeverCommits() {
+        for configuration in [ModelPickerConfiguration.composer, .serverDefault] {
+            XCTAssertNil(
+                ModelPickerSheet.customOption(
+                    modelID: "   ",
+                    providerID: "openai",
+                    configuration: configuration
+                )
+            )
+        }
+    }
+}
